% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{biodbNci-package}
\alias{biodbNci}
\alias{biodbNci-package}
\title{biodbNci: biodbNci, a library for connecting to biodbNci, a library for connecting to the National Cancer Institute (USA) CACTUS Database}
\description{
The biodbNci library is an extension of the biodb framework package. It provides access to biodbNci, a library for connecting to the National Cancer Institute (USA) CACTUS Database. It allows to retrieve entries by their accession number, and run specific web services.
}
\details{
See vignette biodbNci:\preformatted{vignette('biodbNci', package='biodbNci')
}
}
\seealso{
\link{NciCactusConn}.
}
\author{
\strong{Maintainer}: Pierrick Roger \email{pierrick.roger@cea.fr} (\href{https://orcid.org/0000-0001-8177-4873}{ORCID})

}
